refractored code for better UI design

refracted ui queues


maintain code reusability and improve UI


titleEnhancement


removed package-lock.json from individual workspaces


undo formatting of contributing.md


update sidebar ui/ux


fix deprecated package #46


format checker
