+++ backend/.babelrc
@@ -1,10 +1,15 @@
 {
     "presets": ["@babel/preset-env"],
     "plugins": [
+        "@babel/plugin-transform-nullish-coalescing-operator",
+        "@babel/plugin-transform-logical-assignment-operators",
+        "@babel/plugin-transform-json-strings",
+        "@babel/plugin-transform-numeric-separator",
+        "@babel/plugin-transform-optional-chaining",
+        "@babel/plugin-transform-class-properties",
+        "@babel/plugin-transform-export-namespace-from",
         "@babel/plugin-syntax-dynamic-import",
         "@babel/plugin-syntax-import-meta",
-        "@babel/plugin-proposal-class-properties",
-        "@babel/plugin-proposal-json-strings",
         [
             "@babel/plugin-proposal-decorators",
             {
@@ -12,19 +17,14 @@
             }
         ],
         "@babel/plugin-proposal-function-sent",
-        "@babel/plugin-proposal-export-namespace-from",
-        "@babel/plugin-proposal-numeric-separator",
         "@babel/plugin-proposal-throw-expressions",
         "@babel/plugin-proposal-export-default-from",
-        "@babel/plugin-proposal-logical-assignment-operators",
-        "@babel/plugin-proposal-optional-chaining",
         [
             "@babel/plugin-proposal-pipeline-operator",
             {
                 "proposal": "minimal"
             }
         ],
-        "@babel/plugin-proposal-nullish-coalescing-operator",
         "@babel/plugin-proposal-do-expressions",
         "@babel/plugin-proposal-function-bind"
     ]
