+++ backend/package.json
@@ -16,7 +16,7 @@
     "type": "module",
     "dependencies": {
         "@babel/cli": "^7.0.0",
-        "@kubernetes/client-node": "^0.21.0",
+        "@kubernetes/client-node": "^1.0.0",
         "cors": "^2.8.5",
         "dotenv": "^16.4.5",
         "express": "^4.21.0",
@@ -25,26 +25,26 @@
     },
     "devDependencies": {
         "@babel/core": "^7.0.0",
-        "@babel/plugin-proposal-class-properties": "^7.0.0",
         "@babel/plugin-proposal-decorators": "^7.0.0",
         "@babel/plugin-proposal-do-expressions": "^7.0.0",
         "@babel/plugin-proposal-export-default-from": "^7.0.0",
-        "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
         "@babel/plugin-proposal-function-bind": "^7.0.0",
         "@babel/plugin-proposal-function-sent": "^7.0.0",
-        "@babel/plugin-proposal-json-strings": "^7.0.0",
-        "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
-        "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
-        "@babel/plugin-proposal-numeric-separator": "^7.0.0",
-        "@babel/plugin-proposal-optional-chaining": "^7.0.0",
         "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
         "@babel/plugin-proposal-throw-expressions": "^7.0.0",
         "@babel/plugin-syntax-dynamic-import": "^7.0.0",
         "@babel/plugin-syntax-import-meta": "^7.0.0",
+        "@babel/plugin-transform-class-properties": "^7.25.9",
+        "@babel/plugin-transform-export-namespace-from": "^7.25.9",
+        "@babel/plugin-transform-json-strings": "^7.25.9",
+        "@babel/plugin-transform-logical-assignment-operators": "^7.25.9",
+        "@babel/plugin-transform-nullish-coalescing-operator": "^7.26.6",
+        "@babel/plugin-transform-numeric-separator": "^7.25.9",
+        "@babel/plugin-transform-optional-chaining": "^7.25.9",
         "@babel/preset-env": "^7.0.0",
         "@babel/register": "^7.0.0",
         "expect.js": "^0.3.1",
-        "mocha": "^8.0.1",
+        "mocha": "^11.1.0",
         "nodemon": "^3.1.9",
         "sinon": "^7.2.0"
     },
