+++ frontend/src/components/dashboard/DashboardHeader.jsx
@@ -1,0 +1,26 @@
+import React from "react";
+import { Box, IconButton, Tooltip } from "@mui/material";
+import RefreshIcon from "@mui/icons-material/Refresh";
+import TitleComponent from "../Titlecomponent";
+
+const DashboardHeader = ({ onRefresh, refreshing }) => {
+    return (
+        <Box
+            sx={{
+                display: "flex",
+                justifyContent: "space-between",
+                alignItems: "center",
+                mb: 3,
+            }}
+        >
+            <TitleComponent text="Volcano Dashboard" />
+            <Tooltip title="Refresh Data">
+                <IconButton onClick={onRefresh} disabled={refreshing}>
+                    <RefreshIcon />
+                </IconButton>
+            </Tooltip>
+        </Box>
+    );
+};
+
+export default DashboardHeader;
